feature(rooms): Add support for dynamic room descriptions.
This works by replacing the normal description string with a function
that takes the `room` as its first argument, and is expected to return
the description of the room based on the current state of the room and
the game.

This allows great flexibility for content creation.

As an effect, rooms now also know if they've been visited. This allows
room descriptions to be based on if they player have been there yet.