feature(inspect command): Add new `inspect [object]` command.
This will help players learn how a given object can be interacted with.
Even though this is kind of cheating in the puzzles, this will be
helpful if you get stuck. After all, the player makes the choice how
they want to play. Some may want puzzles and others may want a more
straightforward style.